test(integration): fix cart me endpoint test
